Protection for entering the product and variant page by filtering modified request's product_id on url.

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -18,6 +19,9 @@ class VendorProductImageGalleryController extends Controller
     public function index(VendorProductImageGalleryDataTable $dataTable)
     {
         $product = Product::findOrFail(request()->product);
+        if($product->vendor_id != Auth::user()->vendor->id){
+            abort(404);
+        }
         return $dataTable->render('vendor.product.image-gallery.index',compact('product'));
     }
 
@@ -81,7 +85,13 @@ class VendorProductImageGalleryController extends Controller
      */
     public function destroy(string $id)
     {
+        $product = Product::findOrFail(request()->product);
+        if($product->vendor_id != Auth::user()->vendor->id){
+            abort(404);
+        }
+
         $ProductImageGallery = ProductImageGallery::findOrFail($id);
+
         $this->deleteImage($ProductImageGallery->image);
         $ProductImageGallery->delete();
